Initial version of the filescan go module
This supports the scanning of a file for malware via either the default api
key or a registered api key. This module can be installed the usual way by
`go get secplugs.com/filescan`.

The docs have been updated. But might have to go through a couple of iterations
to get it right.